+++ src/main/kotlin/com/example/kbocombo/crawler/infrastructure/KboPlayerListPageParser.kt
@@ -1,4 +1,4 @@
-package com.example.kbocombo.crawler.service
+package com.example.kbocombo.crawler.infrastructure
 
 import com.example.kbocombo.domain.player.vo.PlayerPosition
 import com.example.kbocombo.domain.player.vo.WebId
