+++ src/test/kotlin/com/example/kbocombo/crawler/service/KboPlayerDetailPageParserTest.kt
@@ -1,0 +1,20 @@
+package com.example.kbocombo.crawler.service
+
+import com.example.kbocombo.crawler.infrastructure.KboPlayerDetailPageParser
+import org.junit.jupiter.api.DisplayNameGeneration
+import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores
+import org.junit.jupiter.api.Test
+
+@DisplayNameGeneration(ReplaceUnderscores::class)
+class KboPlayerDetailPageParserTest {
+
+    @Test
+    fun 테스트() {
+        // given
+        val parser = KboPlayerDetailPageParser()
+
+        // when
+
+        // then
+    }
+}
